feat(deps): update Android Command Line Tools to version 13114758
This commit updates the version of the Android SDK Command Line Tools used in the installation scripts for both Linux and macOS to `13114758`.

To make future updates easier, the version number has been extracted into a new `ANDROID_CMD_TOOLS_VERSION` environment variable at the top of the `apply-install-linux` and `apply-install-macos` scripts. This variable is used to construct the download URL dynamically.

Additionally, comments have been added to the scripts, pointing to the official Android Studio page and the repository XML file to make it easier to check for the latest version.